% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prior.R
\name{add_prior_anomaly}
\alias{add_prior_anomaly}
\title{Add a prior on the probability of an anomaly}
\usage{
add_prior_anomaly(priors, prob)
}
\arguments{
\item{priors}{A list containing other, already defined, priors. If NULL, a
list will be started with \code{anomaly} as entry. Else, the provided list will
be extended with \code{anomaly}.}

\item{prob}{The \code{prob} parameter for a Binomial distribution, as in
\code{dbinom()}. The corresponding \code{size} parameter of \code{dbinom} will be the
length of the time series.}
}
\value{
A list that includes a list named \code{anomaly}, which is a list with a
\code{prob} key-value pair.
}
\description{
Add a prior on the probability of an anomaly
}
\examples{
ps <- add_prior_anomaly(priors = list(), prob = 0.01)
print(ps)

# overwrites the existing entry
ps <- add_prior_anomaly(priors = ps, prob = 0.05)
print(ps)

}
\seealso{
\code{\link[=add_prior_trend]{add_prior_trend()}}, \code{\link[=add_prior_level]{add_prior_level()}}, \code{\link[=add_prior_error]{add_prior_error()}},
\code{\link[=add_prior_seasonality]{add_prior_seasonality()}}
}
