% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prior.R
\name{add_prior_error}
\alias{add_prior_error}
\title{Add a prior on the standard deviation of the error component}
\usage{
add_prior_error(shape, rate, priors = NULL, verbose = FALSE)
}
\arguments{
\item{shape}{TBD}

\item{rate}{TBD}

\item{priors}{A list containing other, already defined, priors. If not
provided, a list will be started with \code{error} as entry. Else, the
provided list will be extended with \code{error}.}

\item{verbose}{Should summarizing information be printed using \code{message()}?}
}
\value{
A list that includes a list named \code{error}, which is a list with
\code{shape} and \code{rate} key-value pairs.
}
\description{
Add a prior on the standard deviation of the error component
}
\examples{
ps <- list()
ps <- add_prior_error(3, 2*0.75)
print(ps)

# overwrites the existing entry
ps <- add_prior_error(3, 2*0.5)
print(ps)

}
\seealso{
\code{add_prior_trend()}, \code{add_prior_level()}, \code{add_prior_seasonality()},
\code{add_prior_anomaly()}
}
