% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prior.R
\name{add_prior_seasonality}
\alias{add_prior_seasonality}
\title{Add a prior on the seasonal component}
\usage{
add_prior_seasonality(priors, prob, guess, n, verbose = FALSE, plot = FALSE)
}
\arguments{
\item{priors}{A list containing other, already defined, priors. If NULL, a
list will be started with \code{seasonality} as entry. Else, the provided list
will be extended with \code{seasonality}.}

\item{prob}{Probability that the time series has a seasonality, and that
therefore a seasonal component should be included in the model. This
parameterizes a Bernoulli distribution and should thus be a value in (0,1).}

\item{guess}{Which \eqn{\gamma} parameter do you expect?}

\item{n}{How much weight (n terms of observations) do you want to give your
prior assumption? The larger this value, the more data observations it will
take to wash out your prior distribution in case it is misspecified.}

\item{verbose}{Should summarizing information be printed using \code{message()}?}

\item{plot}{Should a simple plot of the implied Beta density be drawn?}
}
\value{
A list that includes a list named \code{seasonality}, which is a list with
\code{prob}, \code{alpha}, and \code{beta} key-value pairs.
}
\description{
The Beta distribution defined via \code{alpha} and \code{beta} is the prior on the
\eqn{\gamma} parameter of the estimated model. For example, if we expect that
the seasonal component should only update slowly over time and thus only use
a small part of the error component, then a small \code{alpha} and larger \code{beta}
make sense.
}
\details{
The mean of the Beta distribution is \eqn{\mu = \alpha / (\alpha + \beta)}.
If \code{alpha} is smaller or equal than 1, then the density will be highest at 0.
}
\examples{
ps <- add_prior_seasonality(
  priors = NULL, prob = 0.75, n = 6, guess = 1/6, verbose = TRUE, plot = TRUE
)

print(ps)
}
\seealso{
\code{\link[=add_prior_trend]{add_prior_trend()}}, \code{\link[=add_prior_level]{add_prior_level()}}, \code{\link[=add_prior_error]{add_prior_error()}},
\code{\link[=add_prior_anomaly]{add_prior_anomaly()}}
}
