% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prior.R
\name{add_prior_trend}
\alias{add_prior_trend}
\title{Add a prior on the local-linear trend component}
\usage{
add_prior_trend(priors = NULL, prob, guess, n, verbose = FALSE, plot = FALSE)
}
\arguments{
\item{priors}{A list containing other, already defined, priors. If NULL, a
list will be started with \code{trend} as entry. Else, the provided list
will be extended with \code{trend}.}

\item{prob}{Probability that the time series has a local-linear trend, and
that therefore a trend component should be included in the model. This
parameterizes a Bernoulli distribution and should thus be a value in (0,1).}

\item{guess}{Which \eqn{\alpha \cdot \beta} parameter do you expect?}

\item{n}{How much weight (n terms of observations) do you want to give your
prior assumption? The larger this value, the more data observations it will
take to wash out your prior distribution in case it is misspecified.}

\item{verbose}{Should summarizing information be printed using \code{message()}?}

\item{plot}{Should a simple plot of the implied Beta density be drawn?}
}
\value{
A list that includes a list named \code{trend}, which is a list with
\code{prob}, \code{alpha}, and \code{beta} key-value pairs.
}
\description{
The Beta distribution defined via \code{alpha} and \code{beta} is the prior on the
product \eqn{\alpha \cdot \beta}, where \eqn{\alpha} and \eqn{\beta} are the
smoothing parameters for the level and trend components of the model.
}
\details{
The trend component is updated via \eqn{\alpha \cdot \beta \cdot \epsilon_t}.
When \eqn{\alpha=0}, then the trend component will not change over time
because the level component does not adjust to recent observations of the
time series. Every value of \eqn{\beta} then has the same effect.

Specifying the prior directly on the interaction makes it easier to enforce
an expected \emph{effective} behavior of the trend. Independent of the value of
\eqn{\alpha} (except for \eqn{\alpha = 0}), a small value for
\eqn{\alpha \cdot \beta} implies that the trend will adjust slowly to more
recent observations.

The mean of the Beta distribution is \eqn{\mu = \alpha / (\alpha + \beta)}.
If \code{alpha} is smaller or equal than 1, then the density will be highest at 0.
}
\examples{
ps <- add_prior_trend(
  priors = NULL,
  prob = 0.75,
  guess = 1/15,
  n = 15,
  verbose = TRUE,
  plot = TRUE
)

print(ps)
}
\seealso{
\code{\link[=add_prior_seasonality]{add_prior_seasonality()}}, \code{\link[=add_prior_level]{add_prior_level()}}, \code{\link[=add_prior_error]{add_prior_error()}},
\code{\link[=add_prior_anomaly]{add_prior_anomaly()}}
}
