% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tulip.R
\name{assert_priors_error_requirements}
\alias{assert_priors_error_requirements}
\title{Assert that either a prior on the error component is properly specified,
or the initial states are not provided at all such that they are
automatically generated alongside a \code{residuals_global} vector, or
that a suitable \code{residuals_global} vector is provided that can be used
to derive the not provided prior distribution on the error component.}
\usage{
assert_priors_error_requirements(priors, init_states, y)
}
\description{
Assert that either a prior on the error component is properly specified,
or the initial states are not provided at all such that they are
automatically generated alongside a \code{residuals_global} vector, or
that a suitable \code{residuals_global} vector is provided that can be used
to derive the not provided prior distribution on the error component.
}
\keyword{internal}
