% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.heuristika}
\alias{autoplot.heuristika}
\title{Autoplot method for \code{heuristika} objects}
\usage{
\method{autoplot}{heuristika}(
  object,
  ...,
  method = c("components", "fitted")[1],
  date = NULL,
  scales = c("free", "fixed")[1],
  show_anomalies = TRUE,
  show_params = TRUE
)
}
\arguments{
\item{object}{Fitted model object of class \code{heuristika} returned by
\code{\link[=heuristika]{heuristika()}}}

\item{...}{ignored}

\item{method}{One of \code{components} for visualization of the level, trend, seasonal,
and error components of the fitted model (default), or \code{fitted} to
visualize the fitted values in comparison to the input series}

\item{date}{Optional additional vector with dates in format that can be cast
to \code{YYYY-MM-DD} with same length as \code{object$y}, used to create x-axis}

\item{scales}{One of \code{free} or \code{fixed}, passed to the \code{scales} argument of
\code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}; used when \code{method} is \code{"components"}}

\item{show_anomalies}{Logical; when \code{TRUE} (default), observations that were
treated as anomalies during model fit will be marked in orange; used when
\code{method} is \code{"fitted"}}

\item{show_params}{Logical; if \code{TRUE} (default) then fitted params will be
displayed using \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}; used when \code{method} is \code{"fitted"}}
}
\description{
Use \code{ggplot2} to visualize the components or fitted values of a fitted model
of class \code{heuristika}
}
\examples{
set.seed(4278)
y <- rt(100, df = 10) * 10 + 1:100

fitted <- heuristika(y = y, m = 12, family = "norm")

if (requireNamespace("ggplot2")) {
  autoplot(object = fitted, method = "components")
  autoplot(object = fitted, method = "fitted")
}

}
