% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tulip.R
\name{default_object}
\alias{default_object}
\title{Return an empty \code{tulip} object}
\usage{
default_object(y, y_hat, m, method, priors, comment)
}
\arguments{
\item{y}{A time series as numeric vector, may include NAs for some of the
observations}

\item{y_hat}{Numeric vector of fitted values}

\item{m}{The time series' period length in number of observations (e.g., 12
for yearly seasonality with monthly observations, 1 for no seasonality, ...
); does not handle multiple seasonality (e.g. weekly and yearly for daily
observations)}

\item{method}{One of \code{multiplicative} or \code{additive}; see also \code{tulip()}}

\item{priors}{The list of priors provided to \code{tulip()}}

\item{comment}{A string that can be used to describe non-standard return
cases}
}
\description{
Return an empty \code{tulip} object
}
\keyword{internal}
