% Generated by roxygen2 (4.0.1): do not edit by hand
\name{delete_directory}
\alias{delete_directory}
\title{Delete directory}
\usage{
delete_directory(dir, check = TRUE)
}
\arguments{
\item{dir}{a character scalar of the directory to delete}

\item{check}{a logical scalar indicating whether to check with user}
}
\value{
an invisible logical scalar indicating whether successful.
}
\description{
Deletes directory specified by dir after possibly checking with user.
}

