% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_shift.R
\name{detect_shift}
\alias{detect_shift}
\title{Detect shifts from sign to opposite sign in series}
\usage{
detect_shift(trends, window_size, verbose = TRUE)
}
\arguments{
\item{trends}{Series of period-over-period differences describing the trend;
as created in \code{initialize_states()} to compute the initial trend state.}

\item{window_size}{Integer defining the size of the windows used to detect
structural shifts in the time series; since we compare windows based on
the median of the trends in the window, odd values are easier to handle.
\code{window_size} limits how precise the trend shift can be identified. The
shift cannot be identified more precise than within a window of
\code{window_size} candidate values.}

\item{verbose}{Logical; should a warning summarizing a detected shift be
printed?}
}
\description{
Detect shifts from sign to opposite sign in series
}
