% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doy.R
\name{doy}
\alias{doy}
\title{Day of Year}
\usage{
doy(date)
}
\arguments{
\item{date}{object to convert - coerced to class Date.}
}
\value{
An integer vector indicating the day of the year from 
1 - 366.
}
\description{
Calculates day of the year
}
\seealso{
\code{\link{dayte}}
}

