% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_states.R
\name{estimate_season}
\alias{estimate_season}
\title{Get the median value for each seasonal period}
\usage{
estimate_season(x, m, na_fill)
}
\arguments{
\item{x}{Values to be explained by seasonal effects}

\item{m}{The seasonal length}

\item{na_fill}{Value used to replace NAs (usually 1 or 0 or NA)}
}
\value{
Numeric vector of length equal to that of \code{x}
}
\description{
Get the median value for each seasonal period
}
\examples{
x <- rt(n = 55, df = 3) + 5 * cospi((1:55) / 6)
x[seq_along(x) \%\% 12 == 0] <- NA_real_

plot(x, pch = 19)
lines(x)

s <- tulip:::estimate_season(x = x, m = 12, na_fill = 0)
lines(s, col = "blue", lty = 2)

tulip:::estimate_season(x = x[1:7], m = 12, na_fill = 0)
lines(s, col = "blue", lty = 2)

}
\keyword{internal}
