% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit a robust exponential smoothing model via grid search}
\usage{
fit(
  y,
  m,
  family = c("auto", "norm", "cauchy")[1],
  init_states = NULL,
  param_grid = NULL,
  priors = NULL,
  shift_detection = TRUE,
  window_size = 5,
  remove_outliers = TRUE,
  outlier_budget = 5,
  min_obs_for_outlier_sigma = 12,
  seasonality_threshold = 0.5,
  verbose = TRUE
)
}
\description{
Fit a robust exponential smoothing model via grid search
}
