% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_likelihood.R
\name{fit_map}
\alias{fit_map}
\title{Fit error / observation equation and compute loss}
\usage{
fit_map(param_grid, priors, family, m, y, y_hat, n_cleaned)
}
\arguments{
\item{param_grid}{A matrix of possible parameter values used in grid search}

\item{priors}{A named list of lists of prior parameters}

\item{family}{Distribution to be used as likelihood function}

\item{m}{Scalar indicating the suspected seasonality}

\item{y}{Input time series to fit the model against}

\item{y_hat}{Fitted values after fitting the states and parameters to y}

\item{n_cleaned}{Number of values treated as outlier}
}
\description{
Fit error / observation equation and compute loss
}
\examples{
y <- matrix(c(1, 0, 1, -0.05), ncol = 2)
y_hat <- matrix(c(0.9, 0.25, 1.05, 0.05), ncol = 2)
m <- 12
family <- "norm"
param_grid <- initialize_param_grid()[6:7,]
priors <- list(
  error = list(
    shape = 3,
    rate = 2 * 0.75
    # inv_gamma_prior_alpha <- 3
    # inv_gamma_prior_beta <- (inv_gamma_prior_alpha - 1) * 0.75
  ),
  level = list(
    alpha = 2,
    beta = 6
  ),
  trend = list(
    prob = 0.5,
    alpha = 2,
    beta = 18
  ),
  seasonality = list(
    prob = 0.5,
    alpha = 2,
    beta = 6
  ),
  anomaly = list(
    prob = 0.01
  )
)

}
