% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_states_over_grid.R
\name{fit_states_over_grid}
\alias{fit_states_over_grid}
\title{Fit parameters given time series, initial states, and parameter grid}
\usage{
fit_states_over_grid(y, m, init_states, param_grid)
}
\description{
Fit parameters given time series, initial states, and parameter grid
}
