% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowers}
\alias{flowers}
\title{Flowers time series}
\format{
\subsection{\code{flowers}}{

A data frame with 84 rows and 2 columns:
\describe{
\item{date}{The month formatted as date using the first day of the month}
\item{flowers}{The time series observation}
}
}
}
\usage{
flowers
}
\description{
A data frame containing a single synthetic time series with monthly
observations ranging from January 2015 through December 2021.
}
\details{
The values represent some non-negative counts and have anomalies as final two
observations.
}
\examples{

plot(tulip::flowers, type = "l", col = "grey")
points(tulip::flowers, pch = 19, cex = 0.5)

}
\seealso{
\link{flowers_holdout}
}
\keyword{datasets}
