% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowers_holdout}
\alias{flowers_holdout}
\title{Flowers time series (holdout)}
\format{
\subsection{\code{flowers_holdout}}{

A data frame with 12 rows and 2 columns:
\describe{
\item{date}{The month formatted as date using the first day of the month}
\item{flowers}{The time series observation}
}
}
}
\usage{
flowers_holdout
}
\description{
A data frame containing a single synthetic time series with monthly
observations ranging from January 2022 through December 2022. This is a
continuation of the \link{flowers} dataset, generated based on a single data
generating process. Can be used to evaluate models trained on the \link{flowers}
data.
}
\details{
The values represent some non-negative counts and have anomalies as final two
observations.
}
\examples{

plot(rbind(tulip::flowers, tulip::flowers_holdout), col = "white")
abline(v = as.Date("2021-12-15"), lty = 2)
lines(tulip::flowers, type = "l", col = "grey")
lines(tulip::flowers_holdout, type = "l", col = "grey")
points(tulip::flowers, pch = 19, cex = 0.5)
points(tulip::flowers_holdout, pch = 19, cex = 0.5)

}
\seealso{
\link{flowers}
}
\keyword{datasets}
