% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_states.R
\name{get_fitted}
\alias{get_fitted}
\title{Compute the fitted values given \code{level} and \code{seasonal} component,
depending on whether we use \code{additive} or \code{multiplicative} components}
\usage{
get_fitted(level, seasonal, method)
}
\arguments{
\item{level}{A numeric vector representing the level (and trend) component}

\item{seasonal}{A numeric vector representing the seasonal component}

\item{method}{Are we using \code{additive} or \code{multiplicative} components?}
}
\value{
A numeric vector of the same length as \code{level} and \code{seasonal}
}
\description{
Compute the fitted values given \code{level} and \code{seasonal} component,
depending on whether we use \code{additive} or \code{multiplicative} components
}
\keyword{internal}
