% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_states.R
\name{get_residuals}
\alias{get_residuals}
\title{Compute residuals from actuals \code{x} and fitted values \code{fitted}}
\usage{
get_residuals(x, fitted)
}
\arguments{
\item{x}{A numeric vector representing the actuals}

\item{fitted}{A numeric vector representing the fitted values}
}
\value{
A numeric vector of the same length as \code{x} and \code{fitted}
}
\description{
Compute residuals from actuals \code{x} and fitted values \code{fitted}
}
\keyword{internal}
