% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_branches}
\alias{git_branches}
\title{Get Git Branches}
\usage{
git_branches(dir = ".")
}
\arguments{
\item{dir}{string of path to git repository directory}
}
\value{
Character vector of branch names.
}
\description{
Gets git branches
}

