\name{git_checkout}
\alias{git_checkout}
\title{Checkout Git Branch}
\usage{
git_checkout(branch = "master", dir = ".")
}
\arguments{
  \item{branch}{string of branch name}

  \item{dir}{string of path to git repository directory}
}
\value{
Invisible flag indicating whether successful or error.
}
\description{
Changes git branch
}

