% Generated by roxygen2 (4.0.1): do not edit by hand
\name{git_current_branch}
\alias{git_current_branch}
\title{Get Git Current Branch}
\usage{
git_current_branch(dir = ".")
}
\arguments{
\item{dir}{string of path to git repository directory}
}
\value{
String of current branch name.
}
\description{
Gets git current branch
}

