% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/git.R
\name{git_merge}
\alias{git_merge}
\title{Merge Git Branches}
\usage{
git_merge(branch = "dev", dir = ".")
}
\arguments{
\item{branch}{string of branch name}

\item{dir}{string of path to git repository directory}
}
\value{
Invisible flag indicating whether successful or error.
}
\description{
Merges git branch with current branch
}

