% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_pull}
\alias{git_pull}
\title{Pull Git Changes}
\usage{
git_pull(dir = ".", rebase = FALSE)
}
\arguments{
\item{dir}{string of path to git repository directory}

\item{rebase}{flag indicating whether to rebase or merge}
}
\value{
Invisible logical scalar indicating whether successful or error.
}
\description{
Pulls changes from the online git repository.
}

