\name{git_push}
\alias{git_push}
\title{Push Git Commits}
\usage{
git_push(dir = ".")
}
\arguments{
  \item{dir}{string of path to git repository directory}
}
\value{
Invisible logical scalar indicating whether successful or
error.
}
\description{
Pushes commits in git repository in directory \code{dir} to
remote.
}

