% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/git.R
\name{git_push_origin_branch}
\alias{git_push_origin_branch}
\title{Push Origin of Current Git Branch}
\usage{
git_push_origin_branch(branch = "dev", dir = ".")
}
\arguments{
\item{branch}{string of branch name}

\item{dir}{string of path to git repository directory}
}
\value{
Invisible flag indicating whether successful or error.
}
\description{
Pushes origin of git branch
}

