% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/git.R
\name{git_tags}
\alias{git_tags}
\title{Get Git Tags}
\usage{
git_tags(dir = ".")
}
\arguments{
\item{dir}{string of path to git repository directory}
}
\value{
Character vector of tag names
}
\description{
Get Gits Tags
}

