% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tulip.R
\name{guess_error_sd_from_residuals}
\alias{guess_error_sd_from_residuals}
\title{Safe initialization of guess for error prior}
\usage{
guess_error_sd_from_residuals(residuals, y, fallback = 1)
}
\arguments{
\item{residuals}{Residuals of some initial fit, usually returned by
\code{initialize_states()}}

\item{y}{The time series for which we guess the standard deviation of the
residual error}

\item{fallback}{A fallback value that is used when both \code{residuals} and \code{y}
return unusable values}
}
\value{
A strictly positive scalar
}
\description{
Use the MAD of the residuals if possible, else the MAD of the time series \code{y}
else the fallback.
}
\keyword{internal}
