% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_states.R
\name{has_large_seasonal_component}
\alias{has_large_seasonal_component}
\title{Compare residuals to judge whether series has a sizeable seasonal component}
\usage{
has_large_seasonal_component(residuals, residuals_after_seasonality, threshold)
}
\arguments{
\item{residuals}{Residuals from prediction without seasonal component}

\item{residuals_after_seasonality}{Residuals from prediction with seasonal
component}

\item{threshold}{A value between 0 and 1 describing the share of variation
removed from the residuals due to inclusion of the seasonal component}
}
\value{
Logical scalar
}
\description{
Compare residuals to judge whether series has a sizeable seasonal component
}
\examples{
residuals_after_seasonality <- rnorm(n = 50)
residuals <- residuals_after_seasonality + 5 * cospi((1:50) / 6)

plot(residuals, pch = 19)
lines(residuals, lty = 1)
points(residuals_after_seasonality)
lines(residuals_after_seasonality, lty = 2)

tulip:::has_large_seasonal_component(
  residuals = residuals,
  residuals_after_seasonality = residuals_after_seasonality,
  threshold = 0.5
)

}
\keyword{internal}
