% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_params.R
\name{initialize_params_naive}
\alias{initialize_params_naive}
\title{Initialize paramaters to optimize over based on a set of naive models}
\usage{
initialize_params_naive()
}
\value{
A numeric matrix with six named columns: 'alpha', 'one_minus_alpha',
'beta', 'one_minus_beta', 'gamma', 'one_minus_gamma'. The \code{alpha}
paramaters belong to the model's level component, the \code{beta} parameters
to the model's trend component, and the \code{gamma} parameters to the model's
seasonality component. Each pair usually adds up to 1, however dampening
effectively reduces the sum of \code{beta} and \code{one_minus_beta} to less than
1. As per assertions on \code{\link[=tulip]{tulip()}}'s \code{param_grid}, each row must
sum up to a value between 0 and 3, the columns must be named and in
order, and each individual value must be between 0 and 1.
}
\description{
Use this function to generate a parameter grid (matrix) that can be provided
to the \code{param_grid} argument of \code{\link[=tulip]{tulip()}}. The model estimation then
reduces to evaluating all provided combinations and choosing the best one.
See Details for more. Note that there is nothing special about the matrix
generated by this function---you can define a set of possible parameters in
any way that suits you.
}
\examples{
initialize_params_naive()

}
\references{
\describe{
\item{Rob J. Hyndman, Anne B. Koehler, Ralph D. Snyder, and Simone Grose (2002). \emph{A State Space Framework for Automatic Forecasting using Exponential Smoothing Methods}.}{\url{https://doi.org/10.1016/S0169-2070(01)00110-8}}
}
}
\seealso{
\code{\link[=tulip]{tulip()}}, \code{\link[=initialize_params_random]{initialize_params_random()}},
\code{\link[=initialize_params_grid]{initialize_params_grid()}}
}
