% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_states.R
\name{initialize_states_using_previous_fit}
\alias{initialize_states_using_previous_fit}
\title{Use a previous fit's initialization as initialization for a new time series}
\usage{
initialize_states_using_previous_fit(
  object,
  y,
  use_new_residuals_for_anomaly_candidates = FALSE
)
}
\arguments{
\item{object}{The fitted model object returned by \code{\link[=tulip]{tulip()}} of class
\code{tulip}}

\item{y}{A time series as numeric vector, may include NAs for some (but not
all) of the observations}

\item{use_new_residuals_for_anomaly_candidates}{Logical (default \code{FALSE}); if
\code{TRUE}, the residuals of the new time series will be used to judge whether
some observations are anomaly candidates; else only the first
\code{length(object$y)} observations will be used}
}
\description{
This function assumes that the provided \code{y} is a continuation of the time
series used to derive the provided \code{object}. The first \code{length(object$y)}
observations should be equal across \code{y} and \code{object$y}.
}
\keyword{internal}
