% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is_git_remote}
\alias{is_git_remote}
\title{Is Git Remote Repository}
\usage{
is_git_remote(dir = ".")
}
\arguments{
\item{dir}{string of relative or absolute path to directory. Default is current working directory.}
}
\value{
Flag indicating whether directory is a git repository.
}
\description{
Test if directory is a git remote repository.
}

