% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_states.R
\name{make_default_season_state}
\alias{make_default_season_state}
\title{Make a default seasonal state vector}
\usage{
make_default_season_state(n, method)
}
\arguments{
\item{n}{The target length of the state vector}

\item{method}{Are we using an \code{additive} or a \code{multiplicative} seasonal
state component?}
}
\value{
A numeric vector of length \code{n}; all zero if additive seasonal
component, all one if multiplicative seasonal component
}
\description{
Returns a neutral seasonal state vector. Neutral means a vector of 0s for an
additive seasonal component, or a vector of 1s for a multiplicative seasonal
component.
}
\examples{
tulip:::make_default_season_state(n = 14, method = "additive")
tulip:::make_default_season_state(n = 63, method = "multiplicative")

}
\keyword{internal}
