% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_components}
\alias{plot_components}
\title{Plot scaled components of an \code{heuristika} model}
\usage{
plot_components(object, date = NULL, scales = c("free", "fixed")[1])
}
\arguments{
\item{object}{Fitted model object returned by \code{\link[=heuristika]{heuristika()}}}

\item{date}{Optional additional vector with dates in format that can be cast
to \code{YYYY-MM-DD} with same length as \code{object$y}, used to create x-axis}

\item{scales}{One of \code{free} or \code{fixed}, passed to the \code{scales} argument of
\code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}}
}
\description{
This function requires the \link[ggplot2:ggplot2-package]{ggplot2}. Whether its namespace
is available will be checked when the function is run. \code{ggplot2} is only
suggested, not a default import.
}
\examples{
set.seed(4278)
y <- rt(100, df = 10) * 10 + 1:100

fitted <- heuristika(y = y, m = 12, family = "norm")
plot_components(object = fitted)

}
