% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{plot_fitted}
\alias{plot_fitted}
\title{Plot fitted values of an \code{heuristika} model}
\usage{
plot_fitted(object, date = NULL, show_anomalies = TRUE, show_params = TRUE)
}
\arguments{
\item{object}{Fitted model object returned by \code{\link[=heuristika]{heuristika()}}}

\item{date}{Optional additional vector with dates in format that can be cast
to \code{YYYY-MM-DD} with same length as \code{object$y}, used to create x-axis}

\item{show_anomalies}{Logical; when \code{TRUE} (default), observations that were
treated as anomalies during model fit will be marked in orange}

\item{show_params}{Logical; if \code{TRUE} (default) then fitted params will be
displayed using \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}}
}
\description{
This function requires the \link[ggplot2:ggplot2-package]{ggplot2}. Whether its namespace
is available will be checked when the function is run. \code{ggplot2} is only
suggested, not a default import.
}
\examples{
set.seed(4278)
y <- rt(100, df = 10) * 10 + 1:100

fitted <- heuristika(y = y, m = 12, family = "norm")
plot_fitted(object = fitted)

}
