% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot the marginal forecast of a \code{heuristika} model}
\usage{
plot_forecast(
  object,
  paths,
  date = NULL,
  date_future = NULL,
  show_params = TRUE
)
}
\arguments{
\item{object}{Fitted model object returned by \code{\link[=heuristika]{heuristika()}}}

\item{date}{Optional additional vector with dates in format that can be cast
to \code{YYYY-MM-DD} with same length as \code{object$y}, used to create x-axis}

\item{date_future}{Optional additional vector with dates in format that can
be cast to \code{YYYY-MM-DD} with same length as \code{object$y}, used to create
x-axis for forecast paths}

\item{probs}{Numeric vector of probabilities, passed to \code{\link[stats:quantile]{stats::quantile()}}}
}
\description{
This function requires the \link[ggplot2:ggplot2-package]{ggplot2}. Whether its namespace
is available will be checked when the function is run. \code{ggplot2} is only
suggested, not a default import.
}
\examples{
set.seed(4278)
y <- rt(100, df = 10) * 10 + 1:100

fitted <- heuristika(y = y, m = 12, family = "norm")
paths <- draw_paths(object = fitted, h = 12)

plot_forecast(object = fitted, paths = paths)

}
