% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{plot_paths}
\alias{plot_paths}
\title{Plot a few forecast sample paths of a \code{tulip} model}
\usage{
plot_paths(object, date = NULL, date_future = NULL, n = 5, alpha = 0.75)
}
\arguments{
\item{object}{An object of class \code{tulip_paths} as returned by
\code{predict.tulip()}}

\item{date}{Optional additional vector with dates in format that can be cast
to \code{YYYY-MM-DD} with same length as \code{object$y}, used to create x-axis}

\item{date_future}{Optional additional vector with dates in format that can
be cast to \code{YYYY-MM-DD} with same length as \code{object$y}, used to create
x-axis for forecast paths}

\item{n}{Number of paths to add to plot, a small number is recommended to be
able to see the individual paths; scalar integer between 1 and 10}

\item{alpha}{The transparency parameter used when adding the paths to the
plot, provided to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} and \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}}
}
\description{
This function requires the \link[ggplot2:ggplot2-package]{ggplot2}. Whether its
namespace is available will be checked when the function is run. \code{ggplot2} is
only suggested, not a default import.
}
\details{
Note: This function will use \code{\link[base:sample]{base::sample()}} to randomly select paths that
are added to the plot. Set a seed if you require reproducibility.
}
\examples{
set.seed(4278)
y <- rt(100, df = 10) * 10 + 1:100

fitted <- tulip(y = y, m = 12, family = "norm")
paths <- predict(object = fitted, h = 12)

tulip:::plot_paths(object = paths, n = 3)

}
\keyword{internal}
