% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_states.R
\name{repeated_median_line}
\alias{repeated_median_line}
\title{Fit a level and trend to a time series using Repeated Median Regression}
\usage{
repeated_median_line(x)
}
\arguments{
\item{x}{Numeric vector representing a time series}
}
\value{
A list of \code{intercept}, \code{slope}, \code{fitted} values, and \code{residuals}
}
\description{
Fit a level and trend to a time series using Repeated Median Regression
}
\examples{

y <- 1:100 + rnorm(100)
tulip:::repeated_median_line(y)

y <- 1:28 + rt(n = 28, df = 1)
y_idx <- 1:length(y)
lm_fit <- lm(y ~ 1 + y_idx)
rm_fit <- tulip:::repeated_median_line(y)
plot(y, pch = 19, cex = 0.5)
abline(a = 0, b = 1, lty = 2) # ideal fit
abline(
  a = lm_fit$coefficients[1],
  b = lm_fit$coefficients[2],
  col = "darkblue"
)
abline(
  a = rm_fit$intercept,
  b = rm_fit$slope,
  col = "darkorange"
)

y <- as.numeric(RobStatTM::resex)
y_idx <- 1:length(y)
lm_fit <- lm(y ~ 1 + y_idx)
rm_fit <- tulip:::repeated_median_line(y)
plot(y, pch = 19, cex = 0.5)
abline(
  a = lm_fit$coefficients[1],
  b = lm_fit$coefficients[2],
  col = "darkblue"
)
abline(a = rm_fit$intercept, b = rm_fit$slope, col = "darkorange")

}
\references{
https://en.wikipedia.org/wiki/Repeated_median_regression
}
\keyword{internal}
