% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tulip.R
\name{test_prior_error_in_priors}
\alias{test_prior_error_in_priors}
\title{Test the provided prior on error component}
\usage{
test_prior_error_in_priors(priors)
}
\description{
Returns \code{FALSE} if the provided prior on the model's error component does
not fulfill its requirements.
}
\keyword{internal}
