% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tulip.R
\name{test_residuals_global_given_y}
\alias{test_residuals_global_given_y}
\title{Test whether provided residuals vector fits to provided time series}
\usage{
test_residuals_global_given_y(residuals_global, y)
}
\description{
Returns \code{FALSE} if the residuals vector is not of the same length as \code{y},
or not a non-NULL numeric vector.
}
\keyword{internal}
