% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tulip-package.R
\docType{package}
\name{tulip-package}
\alias{tulip}
\alias{tulip-package}
\title{tulip: Robust Probabilistic Forecasts to Tinker With}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An implementation of robust exponential smoothing models that produce probabilistic forecasts using sample paths. Handling of anomalies is integrated into a custom maximum-a-posteriori fitting procedure that exposes most aspects of the model to the user. Initial states can be provided to start training from previous fits, transfer seasonality from other time series. User-defined parameter grids and prior distributions restrict the set of reasonable model choices to robust ones.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://minimizeregret.com}
  \item \url{https://github.com/timradtke/tulip}
  \item Report bugs at \url{https://github.com/timradtke/tulip/issues}
}

}
\author{
\strong{Maintainer}: Tim Radtke \email{tim@timradtke.net}

}
\keyword{internal}
