% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prior.R
\name{add_prior_anomaly}
\alias{add_prior_anomaly}
\title{Add a prior on the probability of an anomaly}
\usage{
add_prior_anomaly(prob, priors = NULL)
}
\arguments{
\item{prob}{The \code{prob} parameter for a Binomial distribution, as in
\code{dbinom()}. The corresponding \code{size} parameter of \code{dbinom} will be the
length of the time series.}

\item{priors}{A list containing other, already defined, priors. If not
provided, a list will be started with \code{anomaly} as entry. Else, the
provided list will be extended with \code{anomaly}.}
}
\value{
A list that includes a list named \code{anomaly}, which is a list with a
\code{prob} key-value pair.
}
\description{
Add a prior on the probability of an anomaly
}
\examples{
ps <- list()
ps <- add_prior_anomaly(prob = 0.01)
print(ps)

# overwrites the existing entry
ps <- add_prior_anomaly(prob = 0.05)
print(ps)

}
\seealso{
\code{add_prior_trend()}, \code{add_prior_level()}, \code{add_prior_error()},
\code{add_prior_seasonality()}
}
