% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prior.R
\name{add_prior_level}
\alias{add_prior_level}
\title{Add a prior on the level component}
\usage{
add_prior_level(alpha, beta, priors = NULL, verbose = FALSE, plot = FALSE)
}
\arguments{
\item{alpha}{The \code{alpha} parameter of a Beta distribution, as \code{shape1} in
\code{dbeta()}. Must be larger than 0.}

\item{beta}{The \code{beta} parameter of a Beta distribution, as \code{shape2} in
\code{dbeta()}. Must be larger than 0.}

\item{priors}{A list containing other, already defined, priors. If not
provided, a list will be started with \code{level} as entry. Else, the
provided list will be extended with \code{level}.}

\item{verbose}{Should summarizing information be printed using \code{message()}?}

\item{plot}{Should a simple plot of the implied Beta density be drawn?}
}
\value{
A list that includes a list named \code{level}, which is a list with
\code{prob}, \code{alpha}, and \code{beta} key-value pairs.
}
\description{
The Beta distribution defined via \code{alpha} and \code{beta} is the prior on the
\eqn{\alpha} parameter of the estimated model. For example, if we expect that
the level of the time series varies slowly over time, and thus only a
small part of the error component is used to update the level at each time
point, then a small \code{alpha} and larger \code{beta} make sense.
}
\details{
Note: In contrast to the trend and seasonal components, we don't specify an
"inclusion" probability \code{prob} on the level component. The level component
is always used.

The mean of the Beta distribution is \eqn{\mu = \alpha / (\alpha + \beta)}.
If \code{alpha} is smaller or equal than 1, than the density will be highest at 0.

Values of \code{alpha} closer to 0 imply a non-fluctating i.i.d. level component,
while values of \code{alpha} closer to 1 imply a more random-walk-like behavior.
}
\examples{
ps <- add_prior_level(
  alpha = 1, beta = 7, verbose = TRUE, plot = TRUE
)

print(ps)

}
\seealso{
\code{add_prior_seasonality()}, \code{add_prior_trend()}, \code{add_prior_error()},
\code{add_prior_anomaly()}
}
