% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prior.R
\name{add_prior_trend}
\alias{add_prior_trend}
\title{Add a prior on the local-linear trend component}
\usage{
add_prior_trend(prob, guess, n, priors = NULL, verbose = FALSE, plot = FALSE)
}
\arguments{
\item{prob}{Probability that the time series has a local-linear trend, and
that therefore a trend component should be included in the model. This
parameterizes a Bernoulli distribution and should thus be a value in (0,1).}

\item{guess}{Which \eqn{\alpha \cdot \beta} parameter do you expect?}

\item{n}{How much weight (n terms of observations) do you want to give your
prior assumption? The larger this value, the more data observations it will
take to wash out your prior distribution in case it is misspecified.}

\item{priors}{A list containing other, already defined, priors. If not
provided, a list will be started with \code{trend} as entry. Else, the
provided list will be extended with \code{trend}.}

\item{verbose}{Should summarizing information be printed using \code{message()}?}

\item{plot}{Should a simple plot of the implied Beta density be drawn?}
}
\value{
A list that includes a list named \code{trend}, which is a list with
\code{prob}, \code{alpha}, and \code{beta} key-value pairs.
}
\description{
The Beta distribution defined via \code{alpha} and \code{beta} is the prior on the
\eqn{\alpha} and \eqn{\beta} of the estimated model. For example, if we
expect that the trend component should only update slowly over time and thus
only use a small part of the error component, then a small \code{alpha} and larger
\code{beta} make sense.
}
\details{
Note: The trend component is updated via
\eqn{\alpha \cdot \beta \cdot \epsilon_t}. The prior on \eqn{\beta} is
implicitly defined via the \code{add_prior_level()} and \code{add_prior_trend()}.

The mean of the Beta distribution is \eqn{\mu = \alpha / (\alpha + \beta)}.
If \code{alpha} is smaller or equal than 1, than the density will be highest at 0.
}
\examples{
ps <- add_prior_trend(
  prob = 0.75, alpha = 1, beta = 14, verbose = TRUE, plot = TRUE
)

print(ps)
}
\seealso{
\code{add_prior_seasonality()}, \code{add_prior_level()}, \code{add_prior_error()},
\code{add_prior_anomaly()}
}
