% Generated by roxygen2 (4.0.1): do not edit by hand
\name{capitalize_first_letter_words}
\alias{capitalize_first_letter_words}
\title{Capitalize First Letter Words}
\usage{
capitalize_first_letter_words(x)
}
\arguments{
\item{x}{character vector}
}
\value{
Modified character vector
}
\description{
Capitalizes first letter of words in character vector.
}

