\name{combine_date_time}
\alias{combine_date_time}
\title{Combine date and time}
\usage{
combine_date_time(date, time)
}
\arguments{
  \item{date}{coerced to as.Date}

  \item{time}{coerced to as.POSIXct}
}
\value{
A datetime object.
}
\description{
Combines date and time into a datetime object.
}

