% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit a robust exponential smoothing model via grid search}
\usage{
fit(
  y,
  m,
  family = c("norm", "cauchy", "nbinom")[1],
  loss = "bic",
  init_states = NULL,
  param_grid = NULL
)
}
\description{
Fit a robust exponential smoothing model via grid search
}
