% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_likelihood.R
\name{fit_likelihood}
\alias{fit_likelihood}
\title{Fit error / observation equation and compute loss}
\usage{
fit_likelihood(y, y_hat, m, family, param_grid)
}
\description{
Fit error / observation equation and compute loss
}
