\name{git_branch}
\alias{git_branch}
\title{Create Git Branch}
\usage{
git_branch(branch = "dev", dir = ".")
}
\arguments{
  \item{branch}{string of branch name}

  \item{dir}{string of path to git repository directory}
}
\value{
Invisible flag indicating whether successful or error.
}
\description{
Create git branch
}

