% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/git.R
\name{git_commit}
\alias{git_commit}
\title{Commit Git Changes}
\usage{
git_commit(message = paste(user(), Sys.time(), sep = ": "), pull = TRUE,
  rebase = FALSE, push = TRUE, dir = ".")
}
\arguments{
\item{message}{string of commit message}

\item{pull}{flag indicating whether to pull after commiting}

\item{rebase}{flag indicating whether to rebase when pulling}

\item{push}{flag indicating whether to push commits}

\item{dir}{string of path to git repository directory}
}
\value{
Invisible logical scalar indicating whether successful or error.
}
\description{
Commits changes to the git repository in directory \code{dir}.
}

