% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/git.R
\name{git_url_http2ssh}
\alias{git_url_http2ssh}
\title{Switch remote URLs from HTTPS to SSH}
\usage{
git_url_http2ssh(dir = ".")
}
\arguments{
\item{dir}{string of path to git repository directory}
}
\value{
Invisible flag indicating whether successful or error.
}
\description{
Switches remote URLs from HTTPS to SSH
}
\seealso{
\url{https://help.github.com/articles/changing-a-remote-s-url}
}

