% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_param_grid.R
\name{initialize_param_grid}
\alias{initialize_param_grid}
\title{Initialize a grid of paramaters to perform grid search}
\usage{
initialize_param_grid()
}
\description{
Currently small for development purposes.
}
