\name{is_git_branch}
\alias{is_git_branch}
\title{Test If Git Branch}
\usage{
is_git_branch(branch = "master", dir = ".")
}
\arguments{
  \item{branch}{string of branch name}

  \item{dir}{string of path to git repository directory}
}
\value{
Flag indicating whether or not branch is a branch
}
\description{
Test whether branch is a git branch
}

