% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tulip}
\alias{package-tulip}
\alias{tulip}
\alias{tulip-package}
\title{Poiscon Consulting's Utility Functions}
\description{
Wrapper functions to query the operating system and manipulate
file extensions. Tulip is a anagram of putil which is short for Poisson
consultings UTILity functions.
}

